+++ tests/LSP6KeyManager/effects/index.ts
@@ -1,0 +1,9 @@
+export * from "./PermissionChangeOwner.test";
+export * from "./PermissionChangeAddPermissions.test";
+export * from "./PermissionCall.test";
+export * from "./PermissionStaticCall.test";
+export * from "./PermissionDelegateCall.test";
+export * from "./PermissionDeploy.test";
+export * from "./PermissionTransferValue.test";
+export * from "./PermissionSign.test";
+export * from "./AllowedAddresses.test";
